Move non-decorated functions to utils file
This will make it easier to read the routes file, which was getting pretty
crowded.  Also, it separates functions by their purpose, like you're supposed
to!  What an idea!
